refactor: use common list data access to users list
