+++ libs/demo/data-access/common/list/src/index.ts
@@ -1,1 +1,3 @@
+export * from './lib/+state/facade';
 export * from './lib/demo-data-access-common-list.module';
+export * from './lib/models/model';
