+++ libs/demo/data-access/common/list/src/lib/demo-data-access-common-list.module.ts
@@ -1,10 +1,10 @@
-import { ModuleWithProviders, NgModule } from '@angular/core';
+import { ModuleWithProviders, NgModule, Type } from '@angular/core';
 import { SharedUtilNotificationModule } from '@demo/shared/util-notification';
 import { LIST_FEATURE_KEY, ListService } from '@ngdux/list';
 import { EffectsModule } from '@ngrx/effects';
-import { ListEffects } from './+state/users/effects';
-import { ListFacade } from './+state/users/facade';
-import { ListReducerManager } from './+state/users/state.service';
+import { ListEffects } from './+state/effects';
+import { ListFacade } from './+state/facade';
+import { ListReducerManager } from './+state/state.service';
 import { DataAccessCommonListModuleConfig, LIST_SERVICE } from './models/model';
 
 @NgModule({
@@ -13,14 +13,14 @@ import { DataAccessCommonListModuleConfig, LIST_SERVICE } from './models/model';
 })
 export class DemoDataAccessCommonListModule {
   static config<T, S = T>(
-    service: ListService<T, S>,
+    service: Type<ListService<T, S>>,
     config: DataAccessCommonListModuleConfig
   ): ModuleWithProviders<DemoDataAccessCommonListModule> {
     return {
       ngModule: DemoDataAccessCommonListModule,
       providers: [
         { provide: LIST_FEATURE_KEY, useValue: config.listFeatureKey },
-        { provide: LIST_SERVICE, useValue: service }
+        { provide: LIST_SERVICE, useClass: service }
       ]
     };
   }
