+++ libs/demo/data-access/common/list/tsconfig.lib.json
@@ -8,5 +8,5 @@
     "types": []
   },
   "exclude": ["src/**/*.spec.ts", "src/test-setup.ts", "jest.config.ts", "src/**/*.test.ts"],
-  "include": ["src/**/*.ts", "src/lib/+state/users/facade.ts"]
+  "include": ["src/**/*.ts", "src/lib/+state/facade.ts"]
 }
