--- libs/demo/data-access/users/src/lib/+state/users/users.effects.spec.ts
@@ -1,8 +1,0 @@
-import { AbstractListEffects } from '@ngdux/list';
-import { UsersEffects } from './users.effects';
-
-describe('UsersEffects', () => {
-  it('shout extend abstract form effects', () => {
-    expect(UsersEffects.prototype).toBeInstanceOf(AbstractListEffects);
-  });
-});
