+++ libs/demo/data-access/users/src/lib/models/users.model.ts
@@ -2,3 +2,5 @@ export interface DataAccessUsersModuleConfig {
   formFeatureKey?: string;
   listFeatureKey?: string;
 }
+
+export const USERS_DEFAULT_FEATURE_KEY = 'users';
